src/main/java/su/medsoft/mis/er/terminal/controllers/DictionaryController.java
package su.medsoft.mis.er.terminal.controllers;

import io.micronaut.http.HttpResponse;
import io.micronaut.http.MediaType;
import io.micronaut.http.annotation.Controller;
import io.micronaut.http.annotation.Get;
import io.micronaut.http.annotation.Produces;
import io.micronaut.security.annotation.Secured;
import io.micronaut.security.rules.SecuredAnnotationRule;
import io.reactivex.Single;
import su.medsoft.mis.er.terminal.responseMessages.DocumentTypesResponseMessage;
import su.medsoft.mis.er.terminal.responseMessages.InsuranceBlankTypesResponseMessage;
import su.medsoft.mis.er.terminal.services.DictionaryServiceImpl;
import su.medsoft.mis.er.terminal.services.interfaces.DictionaryService;

import javax.inject.Inject;

@Secured(SecuredAnnotationRule.IS_AUTHENTICATED)
@Controller("${endpoints.all.path}/dictionary")
public class DictionaryController {

    private final DictionaryService dictionaryService;

    @Inject
    public DictionaryController(DictionaryService dictionaryService) {
        this.dictionaryService = dictionaryService;
    }

    @Get("/document_types")
    @Produces(MediaType.APPLICATION_JSON)
    public HttpResponse<Single<DocumentTypesResponseMessage>> getDocumentTypes() {
        return dictionaryService.getDocumentTypes();
    }

    @Get("/insurance_blank_types")
    @Produces(MediaType.APPLICATION_JSON)
    public HttpResponse<Single<InsuranceBlankTypesResponseMessage>> getInsuranceBlankTypes() {
        return dictionaryService.getInsuranceBlankTypes();
    }
}
